session-2/projects/incident-postmortem/requirements.txt
# Core dependencies - Updated to latest versions (Dec 2024)
langgraph>=1.0.0
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-core>=1.0.0

# Embeddings and vector store
faiss-cpu>=1.9.0

# Data validation
pydantic>=2.10.0

# Utilities
python-dotenv>=1.0.1
rich>=13.9.0  # For nice terminal output
